% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{A}
\alias{A}
\title{A 0-1 adjacency matrix corresponding to the known edges}
\format{An object of class \code{matrix} with 2598 rows and 2598 columns.}
\usage{
A
}
\description{
A 0-1 adjacency matrix corresponding to the known edges
}
\keyword{datasets}
