\name{edgelist}
\alias{edgelist}
\docType{data}
\title{
An example edge list
}
\description{
An example edge list corresponding to known connections in a network.
}
\usage{data(edgelist)}
\format{
  A matrix with two columns, each row defining one edge.
}
\examples{
data(edgelist)
}
\keyword{datasets}
