% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g}
\alias{g}
\title{A directed acyclic graph corresponding to the Adrenergic signaling in cardiomyocytes pathway from KEGG}
\format{An object of class \code{igraph} of length 10.}
\usage{
g
}
\description{
A directed acyclic graph corresponding to the Adrenergic signaling in cardiomyocytes pathway from KEGG
}
\keyword{datasets}
