% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{group}
\alias{group}
\title{The vector of class indicators}
\format{An object of class \code{numeric} of length 520.}
\usage{
group
}
\description{
The vector of class indicators
}
\keyword{datasets}
