\name{netgsa-package}
\alias{netgsa-package}
\alias{netgsa-package}
\docType{package}
\title{Network-Based Gene Set Analysis}
\description{The netgsa-package provides functions for carrying out Network-based Gene Set Analysis by incorporating external information about interactions among genes, as well as novel interactions learned from data.}
\details{
\tabular{ll}{
Package: \tab netgsa\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2015-03-07\cr
License: \tab GPL (>=2)\cr
}
}
\author{Ali Shojaie <ashojaie@uw.edu> and Jing Ma <mjing@umich.edu>}
\references{
Ma, J., Shojaie, A. & Michailidis, G. (2014). Network-based pathway enrichment analysis with incomplete network information, submitted. \url{http://arxiv.org/abs/1411.7919}.

Shojaie, A., & Michailidis, G. (2010). Network enrichment analysis in complex experiments. Statistical applications in genetics and molecular biology, 9(1), Article 22.

Shojaie, A., & Michailidis, G. (2009). Analysis of gene sets based on the underlying regulatory network. Journal of Computational Biology, 16(3), 407-426.
}
\keyword{ package }


\seealso{
\code{\link{glasso}, \link{glmnet}}
}
