\name{nonedgelist}
\alias{nonedgelist}
\docType{data}
\title{
An example for not-an-edge list
}
\description{
An example for not-an-edge list corresponding to the connections that do not exist in a network.
}
\usage{data(nonedgelist)}
\format{
 A matrix with two columns, each row defining one pair of vertices that is not an edge in the network.
}
\examples{
data(nonedgelist)
}
\keyword{datasets}
