% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathways_mat}
\alias{pathways_mat}
\title{Matrix with pathway indicators}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 100 rows and 2598 columns.
}
\usage{
pathways_mat
}
\description{
Matrix with pathway indicators
}
\keyword{datasets}
