% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edgelist}
\alias{edgelist}
\title{A data frame of edges, each row corresponding to one edge}
\format{
An object of class \code{data.frame} with 2959 rows and 4 columns.
}
\usage{
edgelist
}
\description{
A data frame of edges, each row corresponding to one edge
}
\keyword{datasets}
