\name{netgsa-package}
\alias{netgsa-package}
\alias{netgsa-package}
\docType{package}
\title{Network-Based Gene Set Analysis}
\description{The netgsa-package provides functions for carrying out Network-based Gene Set Analysis by incorporating external information about interactions among genes, as well as novel interactions learned from data.}
\details{
\tabular{ll}{
Package: \tab netgsa\cr
Type: \tab Package\cr
Version: \tab 3.1\cr
Date: \tab 2019-03-12\cr
License: \tab GPL (>=2)\cr
}
}
\author{Ali Shojaie <ashojaie@uw.edu> and Jing Ma <jingma@fredhutch.org>}
\references{
Ma, J., Shojaie, A. & Michailidis, G. (2016) Network-based pathway enrichment analysis with incomplete network information. Bioinformatics 32(20):165--3174. \url{https://doi.org/10.1093/bioinformatics/btw410}

Shojaie, A., & Michailidis, G. (2010a). Penalized likelihood methods for estimation of sparse high-dimensional directed acyclic graphs. Biometrika 97(3), 519-538. \url{http://biomet.oxfordjournals.org/content/97/3/519.short}

Shojaie, A., & Michailidis, G. (2010b). Network enrichment analysis in complex experiments. Statistical applications in genetics and molecular biology, 9(1), Article 22. \url{http://www.ncbi.nlm.nih.gov/pubmed/20597848}.

Shojaie, A., & Michailidis, G. (2009). Analysis of gene sets based on the underlying regulatory network. Journal of Computational Biology, 16(3), 407-426. \url{http://www.ncbi.nlm.nih.gov/pmc/articles/PMC3131840/}
}
\keyword{ package }

\seealso{
\code{\link{glmnet}}
}
