% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonedgelist}
\alias{nonedgelist}
\title{A data frame of nonedges, each row corresponding to one negative edge}
\format{An object of class \code{data.frame} with 500 rows and 3 columns.}
\usage{
nonedgelist
}
\description{
A data frame of nonedges, each row corresponding to one negative edge
}
\keyword{datasets}
